fix: Send message 파일 분리 및 로직 짧게 줄이기
- 재활용할 수 있는 로직 재활용
