+++ includes/Server.hpp
@@ -6,7 +6,7 @@
 /*   By: jiychoi <jiychoi@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/29 16:35:50 by san               #+#    #+#             */
-/*   Updated: 2023/01/05 13:55:22 by jiychoi          ###   ########.fr       */
+/*   Updated: 2023/01/05 16:25:01 by jiychoi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,9 +43,8 @@ class Server {
 		void		acceptClient(void);
 
 		void		sendClientMessage(User& user, std::string str);
-		// sender and receiver
+		void		sendClientMessage(User& sender, User& receiver, std::string str);
 		void		sendMessageBroadcast(int mode, Channel& ch, User& sender, std::string str);
-		void		sendMessageUnicast(User& sender, User& receiver, std::string str);
 
 		void		receiveFirstClientMessage(int fdIndex);
 		void		receiveClientMessage(int clientSocket);
